re-factor tests to align with forms file structure
