+++ flow-client/src/test/frontend/form/ValidatorsTests.ts
@@ -4,9 +4,9 @@ const { assert } = intern.getPlugin("chai");
 // API to test
 import { AssertFalse, AssertTrue, DecimalMax, DecimalMin, Digits, Email, Future, Max, Min, Negative, NegativeOrZero, 
   NotBlank, NotEmpty, NotNull, Null, Past, Pattern, Positive, PositiveOrZero, Required, Size } 
-from "../../main/resources/META-INF/resources/frontend/form";
+from "../../../main/resources/META-INF/resources/frontend/form";
 
-  suite("Validator", () => {
+suite("form/Validators", () => {
 
   test("Required", () => {
     const validator = new Required();
